[options] Remove string option type.
As discussed in https://github.com/coq/coq/pull/9876 , "string option"
options are problematic in the wider scope of interacting with other
tools.

This is a historical shell problem, in the sense that for an
environment variable `FOO`, it is hard to distinguish whether `FOO` is
unset or `FOO` is set to `""`.

As noted by @Skyskimmer, Coq only uses this type in a couple of
places, which would be better served anyways by different approaches:

- in the case of proof using, the grammar should capture "unset"
- in the case of the dump, dump filename and dump on / off should be
  different options.