Fix #5752: `Hint Mode` ignored for type classes that appear as assumptions
The creation of the local hint db now inherits the union of the modes
from the dbs passed to `typeclasses eauto`.

We could probably go further and define in a more systematic way the
metadata that should be inherited. A lot of this code could also be
cleaned-up by defining the merge of databases, so that the search code
is parametrized by just one db (the merge of the input ones).

(cherry picked from commit 321d26480444c947ffdaaf849157fd373e40c988)