[meta] [dune] Fix discrepancies in plugin names
We have some discrepancy with the package names for plugins in the
META / Dune case. This PR fixes it.

At some point there was a need for plugin package names having to be
named as their directories.

I think this is not true anymore, but taking the "dir_name ==
package_name" convention just in case.