Document unshelve (#3225)
I believe this is the appropriate place for users to read about this, even tho
`unshelve` is technically a tactical.

This example was explicitly requested in #3225 and is commonly used in both
Iris (and was documented in the changelog at the time).

(cherry picked from commit 9116f7811db0eba707a6d49adcd3d96ca8669c46)