[Classes] Some refactoring on proof save preparation.
We still need to tweak a few things, in particular the conditions on
what is allowed to contain evars, but it is time to see if this works.

Note the ugly bug we have on close_proof:
```
proof_global.ml:278

  let entry_fn p (_, t) =
    let t = EConstr.Unsafe.to_constr t in
    let univstyp, body = make_body t p in
```
which means than arguments of start_proof should be pre-normalized.

This will require a lot of refactoring to be fixed properly.