Tentative alternative fix for #9992.
We crawl the context in the other direction, preventing the allocation of
the return boolean and enhancing sharing. This fast-path is assuming the
heuristic that the variable being renamed always appears in the context,
otherwise we would be renaming for nothing. It seems to always be the
case, due to the way we pick the set of names to be avoided.

(cherry picked from commit eabcd5cfe9d96192747d40776668489f9c5e545c)