Remove outdated CI sanity check from merge script.
This check has been way too frequent and mostly ignored when it was
relevant, so it is completely useless. It was based anyway on a very
shaky heuristics. The only good solution to this problem is a
bors-like workflow, which itself is a work-in-progress.

We don't want annoying useless checks to encourage mergers to bypass checks more often than they should.