Higher level implementation of definition-not-visible warning
This lets us avoid passing ~ontop to do_definition and co, and after #10050
to even more functions.

Instead of checking if there are open proofs when we declare a
variable (which requires that we pass around info about open proofs)
we check after a command that either all proofs are closed or there
are no new variables in the global env.

This seems to work. There is the following strange behaviour, which is
already present in 8.9:
~~~coq
Section foo.

  Lemma foo : Type.
  Proof.
    Variable x : Type. (* x is declared, warn about x *)
    exact nat.
  Qed. (* x is declared, warn about x, foo is defined *)
End foo.
~~~
We don't make things worse in this regard so some other PR can deal
with it.