gitlab: run less jobs unless FULL_CI=true
The idea is that it's unlikely for only 1 of the test suite copies to
fail for a real reason, so we don't need to run all of them.

I would prefer to have only 1 build stage job but later stages depend
on build:base, build:edge+flambda and build:edge+flambda:dune:dev for
non-copy-pasted jobs.

I kept corresponding test suite and validate job copies but they could
also be filtered.