Deprecate "intro_pattern" in tactic notations in favor of "simple_intropattern".
This is to prevent confusion with the terminology used in the grammar
of tactic in the reference manual: "intropattern" in Tactic Notation
and TACTIC EXTEND is actually bound to simple_intropattern and not to
what is called intropattern in the reference manual

After some deprecation phase, "intropattern" could be added back to
mean the "intropattern" of the reference manual.

Note that "simple_intropattern" is actually already available in
"Tactic Notation" with the correct meaning as an entry. Only
"intropattern" is misguiding.

(cherry picked from commit 459b5ce5b2b825db43d645357f83d7fe17289bc5)