[lemmas] Reify info for implicits, univ_decls, prepare for rec_thms.
Key information about an interactive lemma proof was stored as a
closure on an ad-hoc hook, then later made available to the hook
closing actions.

Instead, we put this information in the lemma state and incorporate
these declarations into the normal save path.

We prepare to put the information about rec_thms in the state too.