[rfc] [engine] Remove dependency on `Global` for univgen
This is one of the most impactful changes on the codebase when trying
to make `engine` not to use `Global`.

This patch just showcases the impact, but IMO it is clear that we want
to add the current dirpath to the evar map, or to some other place.

However note that this deserves some thinking, as for example that
would impact `Evd.from_env` etc... so we must discuss how to organize
things.