Refactor and expand contributing guide.
This is almost a full rewrite of the contributing guide.  The goal was
to have a more structured, and more exhaustive guide, where all our
processes are documented, and from which all the useful documentation
is linked to.

The document lists contribution types in the order in which it is most
natural to go through them: from contributions to the ecosystem, to
issues, to code contributions, to taking part in the maintenance.
However, it seemed important to not neglect this last part if we want
to ease the onboarding of new maintainers (and not just of occasional
contributors).

A table of content makes it easy to go through the document (which is
too long to be read from begin to end).

The guide documents our processes in the way they are today, without
changing anything, but this should be a good basis to make them evolve
in the future.  Insufficiently documented tools and processes are
mentioned as such.