Declare section variables in direct style "without" an object
The object was mostly for wrangling universes, but we already have the
universe object for that.

It's also used by some code which iterates over objects to find
variables.

Search used to do this but was changed in a previous commit.

Prettyp.print_context and derivatives do this and I don't understand
it enough to fix it, so I kept a dummy object around. It seems like a
not very common used Print family (not documented AFAICT) so maybe we
should remove it all instead.