[python] Remove use of generic python shebang, update CI
Fixes #10465

Following discussion we don't allow a generic `/usr/bin/python`
shebang anymore. We thus move all the scripts [but one] to python3,
and add python3 to the Azure base environment.

Unfortunately we still depend on python2 for the update-compat.py
script, see #10491

We thus have to complement #10467 adding python2 back to Nix,
until #10491 is fixed.