[funind] Remove export of `generate_functional_principle` and `make_scheme`
This removes the export of two internal functions, moving to their
only use place.

In particular, `make_scheme` was exposing the low-level `proof_entry`
object, which should not do.

This will allow to refactor all these constant-building functions
towards a more uniform use of the API. In particular, all the
functions manipulating proof entries directly are in `Gen_principle`.