Emit Feedback.AddedAxiom in Declare instead of higher layers
This lets us remove the passing around of "status" in comassumption
and generally makes highlighting axiom adding lines in coqide more
reliable as there's no need for per-command code.

If a command adds multiple axioms it will emit AddedAxiom multiple
times, this doesn't seem to be a problem though.

We may wonder if "Fail Fail Axiom" should be highlighted as
"Axiom" (both before and after this commit it is).