[api] Move the tactic engine to its own library.
This is a proof-of-concept PR moving the tactic monad to its own
library.

Indeed, a few design questions do pop-up when doing this move, in
particular on the organization of the generic interpretation code.

While I think that this PR does not feel right yet, it could be a good
moment to discuss the intended design of the infrastructure so we can
advance in a coordinated fashion.