[lemmas] Cleanup users of default proof information.
We remove calls of `Lemmas.Info.make` that where using the default
parameters, as this is mostly dead code now.

This brings into question quite a few things, in particular, the
uneven support of `scope` attributes by different commands / plugins.

We don't attempt to solve that yet, hopefully the ongoing constant
saving path refactoring will be able to take care of these
inconsistencies.