[gitlab/ci] Prevent Corn from running if Bignums has failed.
The needs keyword needs to include all the transitive dependencies.
Otherwise, we risk stage-3 being skipped but stage-4 being run as in
https://gitlab.com/coq/coq/pipelines/78119475.

The reason why we don't need to include bignums in the dependencies of
corn is because the artifact for math-classes is the directory where
both bignums and math-classes were built.