[gitlab/ci] Add explicit jobs for ext-lib and simple-io.
With this commit, the dependency graph declared in GitLab
configuration file reflects the one in Makefile.ci.  The advantage of
this change is that pull request authors and reviewers can more
quickly see which specific project is broken, and what type of project
this is (simple-io and ext-lib, despite being declared as dependencies
of a plugin, are libraries, and thus should get backward-compatible
fixes, and breaking them warrants a changelog entry).

Going further would require reworking the way other external projects
we test manage their dependencies, so this is not on the immediate
roadmap.