Implement a BTT-style restriction for positive coinductive types.
The idea is to check that the return predicate of a coinductive
pattern-matching is stable by cofix expansion. It is literally the
one rule that has to hold for subject reduction to be preserved.

For now this is only performed in the kernel, so as to separate concerns.
We introduce an option to activate the check, but it is not yet on by
default. Whether or not performing the check is decided on a by-case basis
(pun intended), i.e. every coinductive pattern-matching remembers that it
is allowed or not to break subject reduction. Eventually, this backward
compatibility feature will be removed from the kernel, and thus the flag
from the pattern-matching information as well.