Avoid double universe declarations from sections
Before, when declaring a constant in a section, we go
add_constant -> add_constant_aux -> add_field ->
Safe_typing.push_context_set -> Section.push_constraints
then when closing the section we put back the constraints then
add_constant_aux again

Because we fully recheck inductives after discharge (hopefully this
will change in the future) we need all universes declared before an
inductive to be present, but we don't want the universes from the
inductive itself (double declaration). So we have to declare universes
in order with the globals.

To do this add a kind of section_entry for free floating universes
instead of putting them in an unordered blob sec_mono_univs, and avoid
putting universes of globals in there.

We could also try to avoid having add_constant_aux/add_field pushing
universes and declare them separately but this is more difficult.