[declare] Remove declare_scheme hook in Declare
We introduce a new module that registers the scheme information that
side-effects need, thus removing the hook from `Declare`.

As we may want to deprecate scheme side effects, there is no need to
design a general mechanism for this kind of registration for now.

Would we remove the scheme side-effects the scheme code could become
self-contained again.