Implicit arguments: add support for giving implicit arguments by position.
This adds a syntax of the form (1:=t) to make explicit a non-dependent
implicit arguments. In (arg_2:=t), "2" is the position among all
implicit arguments. In (1:=t), "1" is the position among all
non-dependent implicit arguments.

At the same time, we improve the error message when the explicitation
of an implicit argument is ill-named or with a bad index.

Future work:
- For unnamed implicit arguments in inner positions (e.g. in
  "let f `{C} (c:C) := ... in f c")), compute their non-dependent position.
- Deprecate the use of arg_X

Include suggestions co-authored-by: Gaëtan Gilbert <gaetan.gilbert@skyskimmer.net>