Use the Alloc_small macro from the OCaml runtime rather than our own.
We cannot use caml_alloc_small because the macros Setup_for_gc and
Restore_after_gc are still relevant (and critical). This means defining
the CAML_INTERNALS macro, but it is a legit use and actually documented
in the OCaml manual.

This will help with forward compatibility with OCaml compilers, e.g.,
issue #10603. Unfortunately, it also means that we can no longer use #9914
to prevent memory corruption.

The old macro is still used for OCaml versions prior to 4.10, as the
upstream macro might process Ctrl+C when it is called.