Mini-improvements in when to skip coercions or explicitly print implicit args.
If a return type is given to a match/if/let, then we are in context
(and thus may skip coercions or not make explicit those implicit
arguments inferable from context). Note that the notion of "inferable
from context" remains anyway an approximation in the case of implicit
arguments.

The body of a fix/cofix is also in context.

Also fixed an inconsistency with parsing in the scope used to print
the body of a fix.