indTyping: error instead of anomaly for ill-formed template
and do not run template_candidate in the upper layers when the
template attribute is given.

This means we can use an over-approximation in the upper layer
implementation of template_candidate (returning false even in cases
which the kernel may accept) if we ever want to.

(cherry picked from commit 1db8720bf624c202dcc4f1eecdcde803fed4efc2)