Use the relevance flag in CClosure rel contexts in an efficient way.
Rels that exist inside the environment at the time of the closure creation
are fetched in the global environment, while we only use the local list of
relevance for FRels. All the infrastructure was implicitly relying on this
kind of behaviour before the introduction of SProp.

Fixes #11150: pattern is 10x slower in Coq 8.10.0

(cherry picked from commit b8b835cbb7ce5fa12a60184fd83fbde2082d51b3)